feat: 사용자로 부터 자동차를 몇 번 이동 할 것인지를 입력 받아, 위치를 출력시키는 로직 구현
