+++ src/test/java/carracing/PositionTest.java
@@ -1,5 +1,6 @@
 package carracing;
 
+import carracing.domain.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
