+++ src/test/java/carracing/RandomNumberTest.java
@@ -1,5 +1,6 @@
 package carracing;
 
+import carracing.domain.RandomNumber;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
